requirements.txt
fastapi
uvicorn
pika
pymongo
minio
requests
python-dotenv
pydantic-settings
moviepy
imageio-ffmpeg